Rename component files and update references
