+++ src/components/currency-select/currency-select.css
