+++ src/components/currency-select/currency-select.test.tsx
@@ -1,6 +1,6 @@
 import { describe, test, expect, beforeEach } from "vitest";
 import { cleanup, getByText, render, screen } from "@testing-library/react";
-import Currency from "./currency";
+import Currency from "./currency-select";
 import { useState } from "react";
 import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom/vitest";
